Move persistence logic to app storage module
