+++ src/store.js
@@ -8,7 +8,6 @@ import ComputedNotification from './computed/notification';
 import ComputedSetting from './computed/setting';
 import ComputedSeed from './computed/seed';
 import { DEFAULT_ROUTE, DEFAULT_UNIT, DEFAULT_FIAT } from './config';
-import * as log from './action/log';
 
 export class Store {
   constructor() {
@@ -81,44 +80,6 @@ export class Store {
     ComputedNotification(this);
     ComputedSetting(this);
     ComputedSeed(this);
-  }
-
-  async restore(AsyncStorage) {
-    this._AsyncStorage = AsyncStorage;
-    try {
-      const stateString = await this._AsyncStorage.getItem('settings');
-      const state = JSON.parse(stateString);
-      state &&
-        Object.keys(state).forEach(key => {
-          if (typeof this.settings[key] !== 'undefined') {
-            this.settings[key] = state[key];
-          }
-        });
-      log.info('Loaded initial state');
-      this.loaded = true;
-    } catch (err) {
-      log.error('Store load error', err);
-      this.loaded = true;
-    }
-  }
-
-  async save() {
-    try {
-      const state = JSON.stringify(this.settings);
-      await this._AsyncStorage.setItem('settings', state);
-      log.info('Saved state');
-    } catch (error) {
-      log.error('Store save error', error);
-    }
-  }
-
-  async clear() {
-    try {
-      await this._AsyncStorage.clear();
-      log.info('State cleared');
-    } catch (error) {
-      log.error('Store clear error', error);
-    }
   }
 }
